+++ astroquery/ipac/irsa/tests/test_irsa_remote.py
@@ -66,12 +66,19 @@ class TestIrsa:
         # Number of available catalogs may change over time, test only for significant drop.
         # (at the time of writing there are 933 tables in the list).
         assert len(catalogs) > 900
+        assert isinstance(catalogs, dict)
+
+    def test_list_catalogs_filter(self):
+        spitzer_catalogs = Irsa.list_catalogs(filter='spitzer')
+
+        assert len(spitzer_catalogs) == 142
 
     @pytest.mark.parametrize('servicetype', (None, 'sia', 'ssa'))
     def test_list_collections(self, servicetype):
         collections = Irsa.list_collections(servicetype=servicetype)
         # Number of available collections may change over time, test only for significant drop.
         # (at the time of writing there are 104 SIA and 35 SSA collections in the list).
+        assert isinstance(collections, Table)
         if servicetype == 'ssa':
             assert len(collections) > 30
             assert 'sofia_exes' in collections['collection']
@@ -79,6 +86,11 @@ class TestIrsa:
             assert len(collections) > 100
             assert 'spitzer_seip' in collections['collection']
             assert 'wise_allwise' in collections['collection']
+
+    def test_list_collections_filter(self):
+        spitzer_collections = Irsa.list_collections(filter='spitzer')
+
+        assert len(spitzer_collections) == 47
 
     def test_tap(self):
         query = "SELECT TOP 5 ra,dec FROM cosmos2015"